+++ ExpandWorldPrefabs/service/data/DataEntry.cs
@@ -569,6 +569,7 @@ public class DataEntry
 
   private static int Hash(string key)
   {
+    if (Parse.TryInt(key, out var result)) return result;
     if (key.StartsWith("$", StringComparison.InvariantCultureIgnoreCase))
     {
       var hash = ZSyncAnimation.GetHash(key.Substring(1));